Update: Add custom chara
